MapReduce练习：a.wordcount; b.totalcount; c.maxscore; d.IP去重

+++ F_MapReduce01/src/main/java/cn/tedu/wordcount/WordCountReducer.java
@@ -1,5 +1,6 @@
 package cn.tedu.wordcount;
 
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
@@ -12,22 +13,19 @@ import java.io.IOException;
 // KEYOUT, VALUEOUT - 输出的键值类型
 // 当前案例中，最后应该输出每一个字符对应的次数
 public class WordCountReducer
-        extends Reducer<Text, LongWritable, Text, LongWritable> {
+        extends Reducer<Text, IntWritable, Text, IntWritable> {
     // key：键。输入的键
     // values：值。在Reduce开始的时候，会自动的将相同的键对应的值分到一组去
     // 分组完成之后，会为每一个键形成一个迭代器
     // context：环境参数。可以利用这个参数将结果写出
     @Override
-    protected void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
-        // 目的：获取每一个字符对应的次数
-        // key = 'a'
-        // values = {1, 1, 1, 1, 1...}
-        // 记录总次数
-        long sum = 0;
-        // 遍历values来求和
-        for (LongWritable val : values) {
+    protected void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
+        // key是单词
+        // values是次数
+        int sum = 0;
+        for (IntWritable val : values) {
             sum += val.get();
         }
-        context.write(key, new LongWritable(sum));
+        context.write(key, new IntWritable(sum));
     }
 }
